CUtlRBTree: Fix ambiguous overload compiler errors
Ambiguous overload caused by having 2 constructors who's parameters have types that are equal in size. Since these container size types have been explicitly converted to 64bits to align it with the code compiled in the engine, the compiler no longer knows which constructor to call. Added explicit casts to solve the problem. This error was triggered by declaring: static CUtlMap<const char*, int> s_ConVarFlags(DefLessFunc(const char*));

+++ r5dev/public/tier1/utlrbtree.h
@@ -397,7 +397,7 @@ inline CUtlRBTree<T, I, L, M>::CUtlRBTree(int64 growSize, int64 initSize, const 
 
 template < class T, class I, typename L, class M >
 inline CUtlRBTree<T, I, L, M>::CUtlRBTree(const LessFunc_t& lessfunc) :
-	m_Elements(0, 0),
+	m_Elements((int64)0, (int64)0),
 	m_LessFunc(lessfunc),
 	m_Root(InvalidIndex()),
 	m_NumElements(0),
